feat: styled signup page
