+++ src/components/ButtonPrimary/ButtonPrimary.scss
@@ -1,0 +1,19 @@
+@import '../../App.scss';
+
+.button__primary {
+    align-items: center;
+    background-color: $dominant;
+    border-radius: 50px;
+    border: 1px solid $dominant;
+    color: $secondary;
+    cursor: pointer;
+    display: flex;
+    gap: .5em;
+    justify-content: center;
+    padding: .25rem 1rem;
+}
+
+.button__primary:hover {
+    background-color: $accent;
+    border: 1px solid $accent;
+}
