+++ src/components/Input/Input.jsx
@@ -1,4 +1,5 @@
 import PropTypes from 'prop-types';
+
 import './Input.scss';
 
 /**
@@ -15,7 +16,7 @@ export default function Input(props) {
       <label htmlFor={name}>
         {labelText}
       </label>
-      <input type={type} name={name} id={name} onChange={onChange} />
+      <input className="form__input" type={type} name={name} id={name} onChange={onChange} />
       {error && <p className="form__error">{error}</p>}
     </div>
   );
