+++ src/components/Input/Input.scss
@@ -1,0 +1,23 @@
+@import '../../App.scss';
+
+.form__control {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    margin: 10px 0;
+}
+
+.form__input {
+    border-radius: 5px;
+    border: 1px solid $grey;
+    color: $grey;
+    font-family: 'Roboto', sans-serif;
+    padding: 5px 10px;
+    width: 200px;
+}
+
+.form__input:focus {
+    border: 1px solid $accent;
+    outline: 2px solid $accent;
+}
