+++ src/components/Intercept/Intercept.scss
@@ -12,10 +12,21 @@
 }
 
 .intercept__container {
+    align-items: center;
     background-color: white;
-    width: 80%;
-    max-width: 400px;
-    padding: 1em;
     border-radius: 10px;
     box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.5);
+    display: flex;
+    flex-direction: column;
+    gap: .5em;
+    justify-content: center;
+    max-width: 400px;
+    padding: 1em;
+    width: 80%;
 }
+
+.intercept__title, .intercept__body {
+    margin: 0;
+}
+
+
