--- src/pages/Home.jsx
@@ -1,3 +1,0 @@
-export default function Home() {
-  return <h1>Home JSX</h1>;
-}
