+++ src/pages/Home/Home.jsx
@@ -1,0 +1,9 @@
+import './Home.scss';
+
+export default function Home() {
+  return (
+    <div className="home__container">
+      <h1>Home JSX</h1>
+    </div>
+  );
+}
