+++ src/pages/Home/Home.scss
@@ -1,0 +1,5 @@
+@import '../../App.scss';
+
+.home__container {
+    height: $fullScreen;
+}
