+++ src/pages/Signup/Signup.jsx
@@ -1,0 +1,62 @@
+import { useState, useContext } from 'react';
+
+import { Context } from '../../store';
+import { setIntercept } from '../../store/actions';
+
+import { createDocOnEmailSignup } from '../../services/authentication';
+
+import ButtonPrimary from '../../components/ButtonPrimary/ButtonPrimary';
+import GoogleLoginButton from '../../components/GoogleLoginButton/GoogleLoginButton';
+import Input from '../../components/Input/Input';
+import './Signup.scss';
+
+export default function SignUp() {
+  const { dispatch } = useContext(Context);
+  const [form, setForm] = useState({});
+
+  const handlerOnChange = (e) => {
+    setForm({ ...form, [e.target.name]: e.target.value });
+  };
+
+  const handlerEmailSignup = async (e) => {
+    e.preventDefault();
+    if (form.password !== form.confirmPassword) {
+      dispatch(setIntercept({
+        title: 'Error', message: 'Password and confirm password are the same', navigation: '/signup', buttonMsg: 'Try again',
+      }));
+      return null;
+    }
+    const user = await createDocOnEmailSignup(form.email, form.password);
+    if (user.accessToken) {
+      dispatch(setIntercept({
+        title: 'Success',
+        message: 'Sign up successful, please check your inbox or spam folder to verify your email',
+        navigation: '/login',
+        buttonMsg: 'Continue',
+      }));
+    } else {
+      dispatch(setIntercept({
+        title: 'Error', message: 'Sign up failed', navigation: '/signup', buttonMsg: 'Try again',
+      }));
+    }
+    return null;
+  };
+
+  return (
+    <div className="signup-page">
+      <div className="signup__container">
+        <h1 className="signup__title">Sign up</h1>
+        <form className="signup-form" onSubmit={handlerEmailSignup}>
+          <Input type="email" name="email" labelText="Email" onChange={handlerOnChange} />
+          <Input type="password" name="password" labelText="Password" onChange={handlerOnChange} />
+          <Input type="password" name="confirmPassword" labelText="Confirm Password" onChange={handlerOnChange} />
+          <ButtonPrimary isSubmit>Sign up</ButtonPrimary>
+        </form>
+        <GoogleLoginButton isLogin={false}>
+          <img src="/icons/google-icon.svg" alt="Google" className="google__icon" />
+          Sign up with Google
+        </GoogleLoginButton>
+      </div>
+    </div>
+  );
+}
